Folder refactor
