+++ test/group/group_genes.json
@@ -1,0 +1,1 @@
+{"MTOR": ["MTOR"], "ARID1A": ["ARID1A"], "VHL": ["VHL"], "PTEN": ["PTEN"], "TP53": ["TP53"], "tumor_supressors": ["PTEN", "TP53", "VHL"]}
