+++ test/group/group_impacts.json
@@ -1,0 +1,1 @@
+{"missense": ["missense"], "nonsynonymous": ["missense", "nonsense"]}
