+++ test/group/group_samples.json
@@ -1,0 +1,1 @@
+{"K_24_1_A_1": ["K_24_1_A_1"], "K_8_1_A_1": ["K_8_1_A_1"], "K_19_1_A_1": ["K_19_1_A_1"], "K_30_1_A_1": ["K_30_1_A_1"], "K_32_1_A_1": ["K_32_1_A_1"], "K_39_1_A_1": ["K_39_1_A_1"], "K_13_1_A_1": ["K_13_1_A_1"], "K_31_1_A_1": ["K_31_1_A_1"], "K_9_1_A_1": ["K_9_1_A_1"], "K_29_1_A_1": ["K_29_1_A_1"], "all_samples": ["K_17_1_A_1", "K_5_1_A_1", "K_25_1_A_1", "K_27_1_A_1", "K_24_1_A_1", "K_8_1_A_1", "K_19_1_A_1", "K_30_1_A_1", "K_32_1_A_1", "K_39_1_A_1", "K_13_1_A_1", "K_31_1_A_1", "K_9_1_A_1", "K_29_1_A_1", "K_34_1_A_1", "K_6_1_A_1", "K_7_1_A_1", "K_33_1_A_1", "K_38_1_A_1", "K_20_1_A_1", "K_28_1_A_1", "K_21_1_A_1", "K_23_1_A_1", "K_35_1_A_1", "K_18_1_A_1", "K_36_1_A_1", "K_40_1_A_1", "K_16_1_A_1", "K_22_1_A_1", "K_37_1_A_1", "K_26_1_A_1", "K_41_1_A_1", "K_12_1_A_1", "K_15_1_A_1", "K_42_1_A_1", "K_10_1_A_1", "K_14_1_A_1", "K_11_1_A_1"]}
