--- test/group_genes.json
@@ -1,1 +1,0 @@
-{"MTOR": ["MTOR"], "ARID1A": ["ARID1A"], "VHL": ["VHL"], "PTEN": ["PTEN"], "TP53": ["TP53"], "tumor_supressors": ["PTEN", "TP53", "VHL"]}
