--- test/group_impacts.json
@@ -1,1 +1,0 @@
-{"missense": ["missense"], "nonsynonymous": ["missense", "nonsense"]}
