+++ test/kidneypanel4oncodrivefml.bed5.bed
@@ -1,0 +1,152 @@
+CHROMOSOME	START	END	ELEMENT	SEGMENT
+1	11107484	11107500	MTOR	exon58
+1	11108180	11108286	MTOR	exon57
+1	11109289	11109370	MTOR	exon56
+1	11109648	11109729	MTOR	exon55
+1	11112851	11112917	MTOR	exon54
+1	11114317	11114453	MTOR	exon53
+1	11114812	11114887	MTOR	exon52
+1	11115395	11115468	MTOR	exon51
+1	11117003	11117086	MTOR	exon50
+1	11121245	11121368	MTOR	exon49
+1	11121978	11122126	MTOR	exon48
+1	11124497	11124633	MTOR	exon47
+1	11126621	11126796	MTOR	exon46
+1	11127009	11127144	MTOR	exon45
+1	11127623	11127806	MTOR	exon44
+1	11128003	11128126	MTOR	exon43
+1	11128453	11128552	MTOR	exon42
+1	11128854	11128951	MTOR	exon41
+1	11129737	11129838	MTOR	exon40
+1	11130528	11130777	MTOR	exon39
+1	11133079	11133197	MTOR	exon38
+1	11134350	11134466	MTOR	exon37
+1	11139303	11139435	MTOR	exon36
+1	11139532	11139658	MTOR	exon35
+1	11144647	11144755	MTOR	exon34
+1	11144967	11145045	MTOR	exon33
+1	11146675	11146791	MTOR	exon32
+1	11150125	11150226	MTOR	exon31
+1	11157151	11157291	MTOR	exon30
+1	11167441	11167517	MTOR	exon29
+1	11199257	11199403	MTOR	exon28
+1	11199540	11199703	MTOR	exon27
+1	11204560	11204703	MTOR	exon26
+1	11209311	11209458	MTOR	exon25
+1	11210813	11210906	MTOR	exon24
+1	11212311	11212474	MTOR	exon23
+1	11212795	11212908	MTOR	exon22
+1	11213398	11213566	MTOR	exon21
+1	11216147	11216234	MTOR	exon20
+1	11228667	11228700	MTOR	exon
+1	26696403	26697540	ARID1A	exon01
+1	26729650	26729863	ARID1A	exon02
+1	26731151	26731604	ARID1A	exon03
+1	26732675	26732792	ARID1A	exon04
+1	26760855	26761096	ARID1A	exon05
+1	26761383	26761473	ARID1A	exon06
+1	26762151	26762319	ARID1A	exon07
+1	26762972	26763285	ARID1A	exon08
+1	26766220	26766366	ARID1A	exon09
+1	26766456	26766566	ARID1A	exon10
+1	26767789	26767999	ARID1A	exon11
+1	26771118	26771326	ARID1A	exon12
+1	26772499	26772632	ARID1A	exon13
+1	26772811	26772987	ARID1A	exon14
+1	26773345	26773496	ARID1A	exon15
+1	26773579	26773717	ARID1A	exon16
+1	26773801	26773898	ARID1A	exon17
+1	26774328	26775220	ARID1A	exon18
+1	26775576	26775707	ARID1A	exon19
+1	26779022	26780756	ARID1A	exon20
+3	10141847	10142187	VHL	exon01
+3	10146513	10146636	VHL	exon02
+3	10149786	10149965	VHL	exon03
+3	47017092	47017254	SETD2	exon21
+3	47017637	47017739	SETD2	exon20
+3	47019759	47019840	SETD2	exon19
+3	47037665	47037777	SETD2	exon18
+3	47042560	47042700	SETD2	exon17
+3	47046486	47046621	SETD2	exon16
+3	47056820	47057490	SETD2	exon15
+3	47062162	47062346	SETD2	exon14
+3	47067069	47067118	SETD2	exon13
+3	47083719	47084382	SETD2	exon12
+3	47086194	47086314	SETD2	exon11
+3	47088112	47088247	SETD2	exon10
+3	47097954	47098081	SETD2	exon09
+3	47101457	47101555	SETD2	exon08
+3	47103345	47103423	SETD2	exon07
+3	47105996	47106120	SETD2	exon06
+3	47113875	47114004	SETD2	exon05
+3	47116622	47116754	SETD2	exon04
+3	47120181	47124548	SETD2	exon03
+3	47126647	47126663	SETD2	exon02
+3	47163853	47163924	SETD2	exon01
+3	52402287	52402421	BAP1	exon17
+3	52402601	52402674	BAP1	exon16
+3	52402778	52402871	BAP1	exon15
+3	52403137	52403298	BAP1	exon14
+3	52403415	52403894	BAP1	exon13
+3	52404452	52404586	BAP1	exon12
+3	52405109	52405294	BAP1	exon11
+3	52405764	52405912	BAP1	exon10
+3	52406252	52406376	BAP1	exon09
+3	52406828	52406907	BAP1	exon08
+3	52407173	52407316	BAP1	exon07
+3	52407398	52407460	BAP1	exon06
+3	52407957	52408077	BAP1	exon05
+3	52408473	52408606	BAP1	exon04
+3	52409553	52409608	BAP1	exon03
+3	52409713	52409743	BAP1	exon02
+3	52409841	52409878	BAP1	exon01
+3	52548062	52548235	PBRM1	exon30
+3	52550420	52550637	PBRM1	exon29
+3	52550746	52550817	PBRM1	exon28
+3	52554723	52554879	PBRM1	exon27
+3	52561766	52561968	PBRM1	exon26
+3	52563363	52563493	PBRM1	exon25
+3	52564049	52564233	PBRM1	exon24
+3	52576540	52576698	PBRM1	exon23
+3	52579053	52579199	PBRM1	exon22
+3	52586424	52586688	PBRM1	exon21
+3	52587352	52587435	PBRM1	exon20
+3	52589069	52589255	PBRM1	exon19
+3	52603520	52603732	PBRM1	exon18
+3	52609312	52609955	PBRM1	exon17
+3	52615350	52615456	PBRM1	exon16
+3	52617261	52617538	PBRM1	exon15
+3	52627272	52627370	PBRM1	exon14
+3	52628893	52629035	PBRM1	exon13
+3	52634601	52634815	PBRM1	exon12
+3	52641953	52642045	PBRM1	exon11
+3	52643247	52643343	PBRM1	exon10
+3	52644703	52644789	PBRM1	exon09
+3	52648343	52648442	PBRM1	exon08
+3	52651741	52651810	PBRM1	exon07
+3	52658198	52658315	PBRM1	exon06
+3	52662132	52662276	PBRM1	exon05
+3	52668497	52668645	PBRM1	exon04
+3	52678499	52678597	PBRM1	exon03
+3	52679573	52679723	PBRM1	exon02
+3	179218302	179218305	PIK3CA	aaE545
+3	179234295	179234298	PIK3CA	aaH1047
+10	87864469	87864548	PTEN	exon01
+10	87894024	87894109	PTEN	exon02
+10	87925512	87925557	PTEN	exon03
+10	87931045	87931089	PTEN	exon04
+10	87933012	87933251	PTEN	exon05
+10	87952117	87952259	PTEN	exon06
+10	87957852	87958019	PTEN	exon07
+10	87960893	87961118	PTEN	exon08
+10	87965286	87965472	PTEN	exon09
+17	7669608	7669690	TP53	exon11
+17	7670608	7670715	TP53	exon10
+17	7673534	7673608	TP53	exon09
+17	7673700	7673837	TP53	exon08
+17	7674180	7674290	TP53	exon07
+17	7674858	7674971	TP53	exon06
+17	7675052	7675236	TP53	exon05
+17	7675993	7676272	TP53	exon04
+17	7676381	7676403	TP53	exon03
+17	7676520	7676594	TP53	exon02
