+++ test/output/output_mle.tsv
@@ -1,0 +1,133 @@
+gene	sample	impact	dnds	pvalue
+MTOR	K_24_1_A_1	missense	0.38987362	0.02829498
+MTOR	K_24_1_A_1	nonsynonymous	0.42827228	0.034727454
+MTOR	K_8_1_A_1	missense	0.5647296	
+MTOR	K_8_1_A_1	nonsynonymous	0.54189926	
+MTOR	K_19_1_A_1	missense	0.501512	0.075181425
+MTOR	K_19_1_A_1	nonsynonymous	0.5261337	0.07304978
+MTOR	K_30_1_A_1	missense	0.7692831	
+MTOR	K_30_1_A_1	nonsynonymous	0.82117677	
+MTOR	K_32_1_A_1	missense	1.1099674	0.10741013
+MTOR	K_32_1_A_1	nonsynonymous	1.1076272	0.11042309
+MTOR	K_39_1_A_1	missense	0.49242824	
+MTOR	K_39_1_A_1	nonsynonymous	0.5084863	
+MTOR	K_13_1_A_1	missense	0.89709854	
+MTOR	K_13_1_A_1	nonsynonymous	0.8895848	
+MTOR	K_31_1_A_1	missense	0.95044106	
+MTOR	K_31_1_A_1	nonsynonymous	0.9376071	
+MTOR	K_9_1_A_1	missense	0.90460193	
+MTOR	K_9_1_A_1	nonsynonymous	0.87925124	
+MTOR	K_29_1_A_1	missense	0.47282493	0.0005787611
+MTOR	K_29_1_A_1	nonsynonymous	0.45928082	0.0002143383
+MTOR	all_samples	missense	0.9111262	0.10110736
+MTOR	all_samples	nonsynonymous	0.89226913	0.02137196
+ARID1A	K_24_1_A_1	missense	1.1185671	0.099530935
+ARID1A	K_24_1_A_1	nonsynonymous	1.1784986	0.03594774
+ARID1A	K_8_1_A_1	missense	1.2981371	0.0051755905
+ARID1A	K_8_1_A_1	nonsynonymous	1.3705281	0.0014858842
+ARID1A	K_19_1_A_1	missense	0.97838897	
+ARID1A	K_19_1_A_1	nonsynonymous	1.011268	0.63249034
+ARID1A	K_30_1_A_1	missense	0.59855884	
+ARID1A	K_30_1_A_1	nonsynonymous	0.6226032	
+ARID1A	K_32_1_A_1	missense	0.82191706	
+ARID1A	K_32_1_A_1	nonsynonymous	0.8191119	
+ARID1A	K_39_1_A_1	missense	0.583291	0.40354466
+ARID1A	K_39_1_A_1	nonsynonymous	0.6034643	0.54913366
+ARID1A	K_13_1_A_1	missense	1.3000761	0.0018979907
+ARID1A	K_13_1_A_1	nonsynonymous	1.2957488	0.0019515753
+ARID1A	K_31_1_A_1	missense	1.2926766	0.0005992055
+ARID1A	K_31_1_A_1	nonsynonymous	1.3652672	5.954504e-05
+ARID1A	K_9_1_A_1	missense	0.9364351	
+ARID1A	K_9_1_A_1	nonsynonymous	0.89921504	
+ARID1A	K_29_1_A_1	missense	0.86287254	
+ARID1A	K_29_1_A_1	nonsynonymous	0.873698	
+ARID1A	all_samples	missense	1.062444	0.07695848
+ARID1A	all_samples	nonsynonymous	1.0592649	0.085189044
+VHL	K_24_1_A_1	missense	0.23619418	
+VHL	K_24_1_A_1	nonsynonymous	0.289718	
+VHL	K_8_1_A_1	missense	0.23634796	
+VHL	K_8_1_A_1	nonsynonymous	0.29064178	
+VHL	K_19_1_A_1	missense	0.25931156	
+VHL	K_19_1_A_1	nonsynonymous	0.2556463	
+VHL	K_30_1_A_1	missense	0.37016064	
+VHL	K_30_1_A_1	nonsynonymous	0.36972776	
+VHL	K_32_1_A_1	missense	0.67458797	
+VHL	K_32_1_A_1	nonsynonymous	0.67397684	
+VHL	K_39_1_A_1	missense	0.21879286	
+VHL	K_39_1_A_1	nonsynonymous	0.21681252	
+VHL	K_13_1_A_1	missense	0.18916415	0.27457815
+VHL	K_13_1_A_1	nonsynonymous	0.21355543	
+VHL	K_31_1_A_1	missense	0.33161184	
+VHL	K_31_1_A_1	nonsynonymous	0.31973085	
+VHL	K_9_1_A_1	missense	0.6220028	
+VHL	K_9_1_A_1	nonsynonymous	0.6416203	
+VHL	K_29_1_A_1	missense	0.3701433	
+VHL	K_29_1_A_1	nonsynonymous	0.36971366	
+VHL	all_samples	missense	0.9894239	
+VHL	all_samples	nonsynonymous	0.99092305	
+PTEN	K_24_1_A_1	missense	0.66345096	
+PTEN	K_24_1_A_1	nonsynonymous	0.7376958	
+PTEN	K_8_1_A_1	missense	0.1780078	
+PTEN	K_8_1_A_1	nonsynonymous	0.17795625	
+PTEN	K_19_1_A_1	missense	0.7901508	
+PTEN	K_19_1_A_1	nonsynonymous	0.788573	
+PTEN	K_30_1_A_1	missense	0.20243968	0.004174769
+PTEN	K_30_1_A_1	nonsynonymous	0.21724252	0.008500636
+PTEN	K_32_1_A_1	missense	0.6171134	
+PTEN	K_32_1_A_1	nonsynonymous	0.62235415	
+PTEN	K_39_1_A_1	missense	0.3479242	
+PTEN	K_39_1_A_1	nonsynonymous	0.3406627	
+PTEN	K_13_1_A_1	missense	0.6576312	
+PTEN	K_13_1_A_1	nonsynonymous	0.7268872	
+PTEN	K_31_1_A_1	missense	0.89524627	
+PTEN	K_31_1_A_1	nonsynonymous	0.97635067	
+PTEN	K_9_1_A_1	missense	0.7694149	
+PTEN	K_9_1_A_1	nonsynonymous	0.81428003	
+PTEN	K_29_1_A_1	missense	0.30118117	0.04413718
+PTEN	K_29_1_A_1	nonsynonymous	0.28614426	0.014915824
+PTEN	all_samples	missense	1.261946	2.2053719e-06
+PTEN	all_samples	nonsynonymous	1.2804714	2.3841858e-07
+TP53	K_24_1_A_1	missense	0.25557786	
+TP53	K_24_1_A_1	nonsynonymous	0.25077352	
+TP53	K_8_1_A_1	missense	0.24315086	
+TP53	K_8_1_A_1	nonsynonymous	0.24007052	
+TP53	K_19_1_A_1	missense	0.45709062	
+TP53	K_19_1_A_1	nonsynonymous	0.44865933	
+TP53	K_30_1_A_1	missense	0.62628967	
+TP53	K_30_1_A_1	nonsynonymous	0.62066764	
+TP53	K_32_1_A_1	missense	1.0678872	0.23462433
+TP53	K_32_1_A_1	nonsynonymous	1.2102281	0.030183136
+TP53	K_39_1_A_1	missense	0.16883627	0.010859072
+TP53	K_39_1_A_1	nonsynonymous	0.1685249	0.008015573
+TP53	K_13_1_A_1	missense	0.20606574	0.0005825162
+TP53	K_13_1_A_1	nonsynonymous	0.20339026	0.00025880337
+TP53	K_31_1_A_1	missense	0.94823116	
+TP53	K_31_1_A_1	nonsynonymous	1.0480863	0.30816573
+TP53	K_9_1_A_1	missense	1.212694	0.022805333
+TP53	K_9_1_A_1	nonsynonymous	1.3157068	0.0028137565
+TP53	K_29_1_A_1	missense	0.2302131	
+TP53	K_29_1_A_1	nonsynonymous	0.27784652	
+TP53	all_samples	missense	1.2601227	8.881092e-06
+TP53	all_samples	nonsynonymous	1.2668574	3.2782555e-06
+tumor_supressors	K_24_1_A_1	missense	0.6886792	
+tumor_supressors	K_24_1_A_1	nonsynonymous	0.7623039	
+tumor_supressors	K_8_1_A_1	missense	0.417927	
+tumor_supressors	K_8_1_A_1	nonsynonymous	0.47475943	
+tumor_supressors	K_19_1_A_1	missense	1.0113293	0.6319063
+tumor_supressors	K_19_1_A_1	nonsynonymous	0.9889585	
+tumor_supressors	K_30_1_A_1	missense	0.6181591	
+tumor_supressors	K_30_1_A_1	nonsynonymous	0.62338746	
+tumor_supressors	K_32_1_A_1	missense	1.268352	0.006727338
+tumor_supressors	K_32_1_A_1	nonsynonymous	1.296486	0.0034363866
+tumor_supressors	K_39_1_A_1	missense	0.3506133	
+tumor_supressors	K_39_1_A_1	nonsynonymous	0.33600768	
+tumor_supressors	K_13_1_A_1	missense	0.43771407	0.12516427
+tumor_supressors	K_13_1_A_1	nonsynonymous	0.48576778	0.16262746
+tumor_supressors	K_31_1_A_1	missense	1.0265712	0.44719213
+tumor_supressors	K_31_1_A_1	nonsynonymous	1.0967909	0.115342915
+tumor_supressors	K_9_1_A_1	missense	1.1491029	0.03914994
+tumor_supressors	K_9_1_A_1	nonsynonymous	1.2379613	0.005782604
+tumor_supressors	K_29_1_A_1	missense	0.52161694	0.6900891
+tumor_supressors	K_29_1_A_1	nonsynonymous	0.5238647	0.4229247
+tumor_supressors	all_samples	missense	1.2488945	5.9604645e-08
+tumor_supressors	all_samples	nonsynonymous	1.255917	0.0
