--- test/output/results_mle.tsv
@@ -1,131 +1,0 @@
-gene	sample	impact	dnds	pvalue
-MTOR	K_24_1_A_1	Missense	0.52081436	
-MTOR	K_24_1_A_1	NonSynonymous	0.565127	
-MTOR	K_8_1_A_1	Missense	0.63989127	
-MTOR	K_8_1_A_1	NonSynonymous	0.6273867	
-MTOR	K_19_1_A_1	Missense	0.6430056	
-MTOR	K_19_1_A_1	NonSynonymous	0.6621876	
-MTOR	K_30_1_A_1	Missense	1.1358451	0.0793609
-MTOR	K_30_1_A_1	NonSynonymous	1.2545382	0.012565076
-MTOR	K_32_1_A_1	Missense	1.5514077	3.0398369e-05
-MTOR	K_32_1_A_1	NonSynonymous	1.5579162	2.3066998e-05
-MTOR	K_39_1_A_1	Missense	0.62919366	
-MTOR	K_39_1_A_1	NonSynonymous	0.6394916	
-MTOR	K_13_1_A_1	Missense	1.1198684	0.095101535
-MTOR	K_13_1_A_1	NonSynonymous	1.1058074	0.11989784
-MTOR	K_31_1_A_1	Missense	1.0674045	0.19664961
-MTOR	K_31_1_A_1	NonSynonymous	1.0479482	0.2805993
-MTOR	K_9_1_A_1	Missense	1.031042	0.40591973
-MTOR	K_9_1_A_1	NonSynonymous	0.9895937	
-MTOR	K_29_1_A_1	Missense	0.5319213	0.006548047
-MTOR	K_29_1_A_1	NonSynonymous	0.51887643	0.002998352
-MTOR	all_samples	Missense	1.1195551	0.0019538403
-MTOR	all_samples	NonSynonymous	1.0956208	0.011942863
-ARID1A	K_24_1_A_1	Missense	1.1199495	0.09700185
-ARID1A	K_24_1_A_1	NonSynonymous	1.183221	0.03355235
-ARID1A	K_8_1_A_1	Missense	1.2986137	0.00516212
-ARID1A	K_8_1_A_1	NonSynonymous	1.3770655	0.0013539195
-ARID1A	K_19_1_A_1	Missense	0.97524303	
-ARID1A	K_19_1_A_1	NonSynonymous	1.0095356	0.66112995
-ARID1A	K_30_1_A_1	Missense	0.59583086	
-ARID1A	K_30_1_A_1	NonSynonymous	0.61914086	
-ARID1A	K_32_1_A_1	Missense	0.81950194	
-ARID1A	K_32_1_A_1	NonSynonymous	0.8176556	
-ARID1A	K_39_1_A_1	Missense	0.5910991	0.6590816
-ARID1A	K_39_1_A_1	NonSynonymous	0.6102024	
-ARID1A	K_13_1_A_1	Missense	1.3066893	0.0016602278
-ARID1A	K_13_1_A_1	NonSynonymous	1.302404	0.001698494
-ARID1A	K_31_1_A_1	Missense	1.2519764	0.0025132895
-ARID1A	K_31_1_A_1	NonSynonymous	1.2960654	0.0004233718
-ARID1A	K_9_1_A_1	Missense	0.92692345	
-ARID1A	K_9_1_A_1	NonSynonymous	0.89556456	
-ARID1A	K_29_1_A_1	Missense	0.8568187	
-ARID1A	K_29_1_A_1	NonSynonymous	0.8709259	
-ARID1A	all_samples	Missense	1.0711453	0.035486102
-ARID1A	all_samples	NonSynonymous	1.0693771	0.03915733
-VHL	K_24_1_A_1	Missense	0.23620178	
-VHL	K_24_1_A_1	NonSynonymous	0.2901493	
-VHL	K_8_1_A_1	Missense	0.23764002	
-VHL	K_8_1_A_1	NonSynonymous	0.29375356	
-VHL	K_19_1_A_1	Missense	0.25738055	
-VHL	K_19_1_A_1	NonSynonymous	0.25441596	
-VHL	K_30_1_A_1	Missense	0.37012547	
-VHL	K_30_1_A_1	NonSynonymous	0.36973944	
-VHL	K_32_1_A_1	Missense	0.67453396	
-VHL	K_32_1_A_1	NonSynonymous	0.6739913	
-VHL	K_39_1_A_1	Missense	0.21846846	
-VHL	K_39_1_A_1	NonSynonymous	0.21692464	
-VHL	K_13_1_A_1	Missense	0.21592027	
-VHL	K_13_1_A_1	NonSynonymous	0.21386752	
-VHL	K_31_1_A_1	Missense	0.3361298	
-VHL	K_31_1_A_1	NonSynonymous	0.32650372	
-VHL	K_9_1_A_1	Missense	0.62368613	
-VHL	K_9_1_A_1	NonSynonymous	0.6455504	
-VHL	K_29_1_A_1	Missense	0.37010556	
-VHL	K_29_1_A_1	NonSynonymous	0.3697245	
-VHL	all_samples	Missense	1.0114391	0.6182163
-VHL	all_samples	NonSynonymous	1.0183067	0.52296066
-PTEN	K_24_1_A_1	Missense	0.6681073	
-PTEN	K_24_1_A_1	NonSynonymous	0.74790454	
-PTEN	K_8_1_A_1	Missense	0.17800334	
-PTEN	K_8_1_A_1	NonSynonymous	0.1779533	
-PTEN	K_19_1_A_1	Missense	0.7901083	
-PTEN	K_19_1_A_1	NonSynonymous	0.78854877	
-PTEN	K_30_1_A_1	Missense	0.22328052	0.026980877
-PTEN	K_30_1_A_1	NonSynonymous	0.21872412	0.012051523
-PTEN	K_32_1_A_1	Missense	0.6213367	
-PTEN	K_32_1_A_1	NonSynonymous	0.6262338	
-PTEN	K_39_1_A_1	Missense	0.3521204	
-PTEN	K_39_1_A_1	NonSynonymous	0.34480178	
-PTEN	K_13_1_A_1	Missense	0.65903884	
-PTEN	K_13_1_A_1	NonSynonymous	0.7314314	
-PTEN	K_31_1_A_1	Missense	0.92199373	
-PTEN	K_31_1_A_1	NonSynonymous	1.0093964	0.66362035
-PTEN	K_9_1_A_1	Missense	0.8009681	
-PTEN	K_9_1_A_1	NonSynonymous	0.83879536	
-PTEN	K_29_1_A_1	Missense	0.3051067	0.058355927
-PTEN	K_29_1_A_1	NonSynonymous	0.2897313	0.020442247
-PTEN	all_samples	Missense	1.2946371	1.7881393e-07
-PTEN	all_samples	NonSynonymous	1.3185837	0.0
-TP53	K_24_1_A_1	Missense	0.22739738	0.81414515
-TP53	K_24_1_A_1	NonSynonymous	0.22457452	0.41599095
-TP53	K_8_1_A_1	Missense	0.2140089	
-TP53	K_8_1_A_1	NonSynonymous	0.21238026	
-TP53	K_19_1_A_1	Missense	0.4653218	
-TP53	K_19_1_A_1	NonSynonymous	0.4575507	
-TP53	K_30_1_A_1	Missense	0.5719785	
-TP53	K_30_1_A_1	NonSynonymous	0.5627144	
-TP53	K_32_1_A_1	Missense	1.0052075	0.746647
-TP53	K_32_1_A_1	NonSynonymous	1.1338427	0.086999655
-TP53	K_13_1_A_1	Missense	0.19425806	0.0002862811
-TP53	K_13_1_A_1	NonSynonymous	0.19238853	0.00013321638
-TP53	K_31_1_A_1	Missense	0.9412928	
-TP53	K_31_1_A_1	NonSynonymous	1.0530429	0.28480124
-TP53	K_9_1_A_1	Missense	1.0699795	0.21560723
-TP53	K_9_1_A_1	NonSynonymous	1.2071931	0.02326262
-TP53	K_29_1_A_1	Missense	0.19625784	
-TP53	K_29_1_A_1	NonSynonymous	0.22923248	
-TP53	all_samples	Missense	1.2078652	0.0006348491
-TP53	all_samples	NonSynonymous	1.226102	0.00019955635
-tumor_supressors	K_24_1_A_1	Missense	0.6677933	
-tumor_supressors	K_24_1_A_1	NonSynonymous	0.740834	
-tumor_supressors	K_8_1_A_1	Missense	0.37072793	
-tumor_supressors	K_8_1_A_1	NonSynonymous	0.43007225	
-tumor_supressors	K_19_1_A_1	Missense	1.0128534	0.6094433
-tumor_supressors	K_19_1_A_1	NonSynonymous	0.9899697	
-tumor_supressors	K_30_1_A_1	Missense	0.62501717	
-tumor_supressors	K_30_1_A_1	NonSynonymous	0.6078905	
-tumor_supressors	K_32_1_A_1	Missense	1.2607392	0.00820595
-tumor_supressors	K_32_1_A_1	NonSynonymous	1.2874581	0.0042586327
-tumor_supressors	K_39_1_A_1	Missense	0.31497985	
-tumor_supressors	K_39_1_A_1	NonSynonymous	0.30460754	0.6334628
-tumor_supressors	K_13_1_A_1	Missense	0.45878792	0.22001755
-tumor_supressors	K_13_1_A_1	NonSynonymous	0.47155827	0.18432957
-tumor_supressors	K_31_1_A_1	Missense	1.0476578	0.2962489
-tumor_supressors	K_31_1_A_1	NonSynonymous	1.1227238	0.06748831
-tumor_supressors	K_9_1_A_1	Missense	1.123351	0.06926882
-tumor_supressors	K_9_1_A_1	NonSynonymous	1.2134063	0.010691404
-tumor_supressors	K_29_1_A_1	Missense	0.4903088	0.5072944
-tumor_supressors	K_29_1_A_1	NonSynonymous	0.49551073	0.33905894
-tumor_supressors	all_samples	Missense	1.2453625	1.1920929e-07
-tumor_supressors	all_samples	NonSynonymous	1.2531335	0.0
