+++ test/step2_estimate.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+
+python ../src/estimator/main.py --option bayes --cores 2 input_estimator.json ./output/output_bayes.tsv
